Update download style
