+++ src/webui/src/static/style/overview.scss
@@ -49,16 +49,3 @@
     }
 }
 
-Button.downStyle{
-    height: 22px;
-}
-
-.downloadLog{
-    width: 9px;
-}
-
-.downMargin{
-    margin-left: 4px;
-    margin-right: 4px;
-}
-
